requirements.txt
nextcord.py==2.5.0